feat: add checks to hide "unwanted files" errors on files in .gitignore
