+++ src/extension.js
@@ -1,35 +1,17 @@
 const vscode = require('vscode');
-const cp = require('child_process');
 const fs = require('fs');
-const {getCurrentDateAsString, getSeverityAsVSCode, getReportErrorRange} = require('./utils.js');
+const {getSeverityAsVSCode, getReportErrorRange, fileIsInGitignore} = require('./utils.js');
+const {startCommand} = require('./system.js');
 // @ts-ignore
 const codingStyleExtractData = require('../epitech_c_coding_style.json')
 
 const bananaDiagnosticCollection = vscode.languages.createDiagnosticCollection('banana-coding-style');
 
-const DOCKER_IMAGE_URL = 'ghcr.io/epitech/coding-style-checker:latest'
-
-/**
- * Starts a command and returns a promise that resolves when the command is done.
- * @param {string} command 
- * @returns {Promise<>}
- */
-async function startCommand(command) {
-	return new Promise((resolve, reject) => {
-		cp.exec(command, (err, stdout, stderr) => {
-			if (err) {
-				console.error(stderr)
-				reject(err);
-			} else {
-				resolve(stdout);
-			}
-		});
-	});
-}
+const DOCKER_IMAGE_URL = 'ghcr.io/epitech/coding-style-checker:latest';
 
 /**
  * Reads a file and returns all its lines as an array.
- * @param {*} file 
+ * @param {string} file 
  * @returns {string[]}
  */
 function getAllFilesLines(file) {
@@ -97,6 +79,10 @@ function processReportError(file, line, severity, language, subgroup, projectFil
 		}
 	}
 
+	if (subgroup === "O1" && fileIsInGitignore(file)) {
+		return;
+	}
+
 	let diagnostic = new vscode.Diagnostic(range, message, severityVSC);
 	diagnostic.code = fullCode;
 	diagnostic.source = 'Banana';
@@ -109,8 +95,8 @@ function processReportError(file, line, severity, language, subgroup, projectFil
 
 /**
  * Processes the report file to display results in the editor.
- * @param {*} reportFile The report file generated by Banana.
- * @param {*} projectFileMap The project file map generated for the current project.
+ * @param {Buffer} reportFile The report file generated by Banana.
+ * @param {Map} projectFileMap The project file map generated for the current project.
  */
 function processReportFile(reportFile, projectFileMap) {
 	const deliveryDir = `${vscode.workspace.workspaceFolders[0].uri.path}`;
@@ -119,6 +105,7 @@ function processReportFile(reportFile, projectFileMap) {
 	var match;
 
 	while ((match = errorRegex.exec(fileContent)) !== null) {
+		// eslint-disable-next-line no-unused-vars
 		let [_, file, lineStr, severity, language, subgroup] = match;
 		let line = parseInt(lineStr);
 		let projectFile = projectFileMap.get(file);
@@ -130,8 +117,8 @@ function processReportFile(reportFile, projectFileMap) {
 
 /**
  * Processes all the reports files generated by Banana.
- * @param {*} reportsDir 
- * @param {*} projectFiles 
+ * @param {string} reportsDir 
+ * @param {Array<Object>} projectFiles 
  */
 function processAllReportFiles(reportsDir, projectFiles) {
 	var projectFileMap = new Map();
@@ -179,7 +166,7 @@ function startBanana() {
 		
 		bananaDiagnosticCollection.clear();
 		processAllReportFiles(reportsDir, projectFiles);
-		await startCommand(`rm -rfd "${reportsDir}"`);
+		await startCommand(`rm -rfd "${reportsDir}" && rmdir "${reportsDir}"`);
 	});
 }
